arm64: dts: mediatek: mt7986: fix the switch reset line on BPI-R3
The current GPIO definition for the MT7531 switch reset line incorrectly specifies a pin GPIO_0 (GPIO5) that is connected to the boot mode selection DIP switch (SW1).

In the public schematic for the BPI-R3 (titled "BPI-R3-MT7986A", revision "V1.1", sheet 4), the SPI1_CS (GPIO32) pin of the MT7986 SoC (U1E) is connected to the GbE_RESET net.

Frank Wunderlich has told me, via the Banana Pi forum, that this GbE_RESET net connects to the MT7531 reset line in the private part of the shematic he has a copy of.