arm64: dts: mediatek: mt7986: include the missing UART1 CTS/RTS pins in BPI-R3
The current definition of the uart1_pins pin group does not include the CTS and RTS pins that are available on header CON1.

In the public schematic for the BPI-R3 (titled "BPI-R3-MT7986A", revision "V1.1", sheet 4), the UART1_RTS and UART1_CTS pins are connected via nets to the connector CON1 (sheet 14) UART0-RTS and UART0-CTS pins. The datasheet does not show these nets as connected to anything else.

These pins can be configured for either UART1 or as a GPIO, with no other alternate usage listed in the pinctrl driver. By changing the pin group definition from uart1_rx_tx to uart1 the additional RTS and CTS pins are included.