arm64: dts: mediatek: mt7986: add missing USB port regulators to BPI-R3
The ssusb node for the USB port is missing the required vbus-supply and vusb33-supply regulators. This results in a log warning when the XHCI device is probed:
xhci-mtk 11200000.usb: supply vbus not found, using dummy regulator
xhci-mtk 11200000.usb: supply vusb33 not found, using dummy regulator

To fix the warning, this patch: connects the existing 3.3V fixed regulator to the vusb33-supply, and adds a new 5V fixed regulator for the vbus-supply. The board has dedicated regulators for the USB ports, but these are always enabled, and can be modelled using fixed regulators.