arm64: dts: mediatek: mt7986: add missing SNFI bus to BPI-R3
The BPI-R3 has an empty pad (labelled U14) for a SNFI flash chip which can be populated.

In the public schematic for the BPI-R3, revision V1.1, sheet 4, the SNFI-NAND flash chip (U14) is connected to the MT7986 SoC (U1E) SNFI/SPI0 bus pins.

The snand device tree node for the SNFI bus is documented in the public MediaTek SDK OpenWrt feed. I have copied this code, but changed the clock definitions to match the names in dt-bindings/clock/mt7986-clk.h

The snand node and associated pin group is also added to the BPI-R3, but left disabled to reflect the factory configuration of the board.

I have not been able to test the SNFI interface yet as I will need to solder a suitable chip in. I am not sure what the correct clock and interrupt values for the ECC are.

Link: https://drive.google.com/file/d/1ED2z_e01Y0e6LDzMC_vLhCsTHEZTGwcg/view?usp=sharing
Link: https://git01.mediatek.com/plugins/gitiles/openwrt/feeds/mtk-openwrt-feeds/+/refs/heads/master/21.02/files/target/linux/mediatek/files-5.4/arch/arm64/boot/dts/mediatek/mt7986a.dtsi#527