arm64: dts: mediatek: mt7986: add missing audio codec interface to BPI-R3
The MT7986 SoC has an audio interface which is supported by sound/soc/mediatek/mt7986/mt7986-afe-pcm.c. However the device tree node is currently missing, so I have copied the code from OpenWrt to add the missing node.

The audio interface is exposed on the header CON1 on the BPI-R3. In the public schematic for the BPI-R3, revision V1.1, sheet 14, the MT7986 SoC PCM_FS, PCM_CLK, PCM_DTX and PCM_DRX pins are connected via nets to the connector CON1 PCM-FS, PCM-BITCLK, PCM-TX and PCM-RX/SPI-CE1 pins.

The audio interface is added to the BPI-R3 device tree, but left in a disabled state to avoid potential pin conflict with the use of the CON1 PCM-RX/SPI-CE1 as a SPI chip enable.

A user can add a suitable device tree overlay to enable the interface along adding an audio codec such as the WM8960 chip. There is an example of such an overlay in OpenWrt.

Link: https://drive.google.com/file/d/1ED2z_e01Y0e6LDzMC_vLhCsTHEZTGwcg/view?usp=sharing
Link: https://git01.mediatek.com/plugins/gitiles/openwrt/feeds/mtk-openwrt-feeds/+/2989c4c3eab839d2d87a18a9bb0cf76d029c927a/21.02/files/target/linux/mediatek/files-5.4/arch/arm64/boot/dts/mediatek/mt7986a.dtsi#527
Link: https://git.openwrt.org/?p=openwrt/openwrt.git;a=blob;f=target/linux/mediatek/patches-6.6/863-arm64-dts-mt7986-add-sound-wm8960.patch;h=15e30dec56cf426757d3585654f8dd550cc2b0b4;hb=166d28ffeaa87de67e404258befbe56a4375ac97